+++ packages/datadog_flags/lib/datadog_flags.dart
@@ -13,6 +13,7 @@ library;
 export 'src/datadog_flags_config.dart'
     show DatadogFlagsConfig, DatadogFlagsSite;
 export 'src/datadog_flags.dart' show DatadogFlags;
+export 'src/datadog_openfeature_provider.dart' show DatadogOpenFeatureProvider;
 export 'src/flags_client.dart'
     show
         DatadogFlagsClient,