+++ tools/third_party_scanner/bin/third_party_scanner.dart
@@ -14,7 +14,6 @@ final projectList = [
   '$root/packages/datadog_common_test',
   '$root/packages/datadog_flags',
   '$root/packages/datadog_flags_flutter',
-  '$root/packages/datadog_openfeature_provider',
   '$root/packages/datadog_flutter_plugin',
   '$root/packages/datadog_flutter_plugin/example',
   '$root/packages/datadog_flutter_plugin/e2e_test_app',
@@ -32,7 +31,6 @@ final projectList = [
 final ignorePackages = [
   "datadog_flags",
   "datadog_flags_flutter",
-  "datadog_openfeature_provider",
   "datadog_flutter_plugin",
 ];